AV-2561 support attach uploads for maps
